requirements.txt
python-dotenv
pyyaml
pandas
facebook_business
streamlit
